Stop inserting spaces
I don't recall why I started inserting spaces before fractions, but it
leads to extra spaces at the beginning of a line or when a fraction
occurs after a word. Stop doing it, and also remove white space
preceding a fraction when it occurs at the beginning of a line.